Split up notebooks by Figures!
Let me know if this makes sense to you all @SoojungYang @yehlincho, but basically I'm hoping to organize our work around one frozen model (mdm2_bilstm_best.h5), each notebook is appended with which figure it corresponds to. F0 in this case is kind of a catch all for the supplemental figures / anything we don't yet have a place for.

Hope it all makes sense!